# gateway/swagger/swagger.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Microservices Gateway API",
    "description": "HTTP Gateway для микросервисной архитектуры",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/github.com.krus210.balun_microservices.protobuf.auth.v1.proto.AuthService/Login": {
      "post": {
        "summary": "Login - Аутентификация пользователя",
        "operationId": "AuthService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.auth.v1.proto.AuthService/Refresh": {
      "post": {
        "summary": "Refresh - Запрос нового токена при протухании текущего",
        "operationId": "AuthService_Refresh",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoRefreshResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoRefreshRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.auth.v1.proto.AuthService/Register": {
      "post": {
        "summary": "Register - Регистрация пользователя",
        "operationId": "AuthService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoRegisterRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protoLoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "title": "email -электронная почта"
        },
        "password": {
          "type": "string",
          "title": "password - пароль пользователя"
        }
      },
      "title": "LoginRequest - запрос Login"
    },
    "protoLoginResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "title": "accessToken - токен доступа"
        },
        "refreshToken": {
          "type": "string",
          "title": "refreshToken - токен после протухания accessToken"
        },
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - уникальный идентификатор пользователя"
        }
      },
      "title": "LoginResponse - ответ Login"
    },
...
    "protoRefreshRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string",
          "title": "refreshToken - токен после протухания accessToken"
        }
      },
      "title": "RefreshRequest - запрос Refresh"
    },
    "protoRefreshResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "title": "accessToken - токен доступа"
        },
        "refreshToken": {
          "type": "string",
          "title": "refreshToken - токен после протухания accessToken"
        },
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - уникальный идентификатор пользователя"
        }
      },
      "title": "RefreshResponse - ответ Refresh"
    },
    "protoRegisterRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "title": "email -электронная почта"
        },
        "password": {
          "type": "string",
          "title": "password - пароль пользователя"
        }
      },
      "title": "RegisterRequest - запрос Register"
    },
    "protoRegisterResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - уникальный идентификатор пользователя"
        }
      },
      "title": "RegisterResponse - ответ Register"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}